Adding initial e2e tests for catalogue items site
Adds a record exporter combining various record/item formats and a prototype site exporter combining record exports with static resources.

Existing exporters adjusted. Needed to allow a static site to be generated for use in e2e tests. Fixtures for generating the site and serving as a local site added for use in tests.

Adds a minimal product item for use in tests.