Removing data catalogue logic
The catalogue is now a standalone project (in an intermediate state).

As the core problem with the performance of publishing records has been resolved, this project now only needs to export records in JSON format for importing into the catalogue as an out of scope, and routine, activity.

The new catalogue is imported as a dependency for access to the metadata library data classes and item base classs/enums.